exa: factor out memory management
Factor out memory management to ease navigation through the code.
